relocate classes
