+++ common/src/main/java/us/myles/ViaVersion/api/type/types/version/Types1_13_2.java
@@ -2,6 +2,8 @@ package us.myles.ViaVersion.api.type.types.version;
 
 import us.myles.ViaVersion.api.minecraft.metadata.Metadata;
 import us.myles.ViaVersion.api.type.Type;
+import us.myles.ViaVersion.api.type.types.Particle;
+import us.myles.ViaVersion.protocols.protocol1_13_2to1_13_1.types.Particle1_13_2Type;
 
 import java.util.List;
 
@@ -15,4 +17,9 @@ public class Types1_13_2 {
      * Metadata type for 1.13
      */
     public static final Type<Metadata> METADATA = new Metadata1_13_2Type();
+
+    /**
+     * Particle type for 1.13.2
+     */
+    public static Type<Particle> PARTICLE = new Particle1_13_2Type();
 }
