+++ common/src/main/java/us/myles/ViaVersion/api/type/types/version/Types1_14.java
@@ -2,8 +2,8 @@ package us.myles.ViaVersion.api.type.types.version;
 
 import us.myles.ViaVersion.api.minecraft.metadata.Metadata;
 import us.myles.ViaVersion.api.type.Type;
+import us.myles.ViaVersion.api.type.types.Particle;
 import us.myles.ViaVersion.api.type.types.minecraft.Particle1_14Type;
-import us.myles.ViaVersion.protocols.protocol1_13to1_12_2.data.Particle;
 
 import java.util.List;
